backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx==0.27.2
aiosqlite==0.20.0
beautifulsoup4==4.12.3
lxml==5.3.0
pyyaml==6.0.2
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
petrovich==1.0.3
rapidfuzz==3.10.1
phonenumbers==8.13.49
openpyxl==3.1.5
structlog==24.4.0
playwright==1.47.0
websockets==13.1
